moving coingecko file
